pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datafaker"
version = "0.1.0"
description = "Multi-threaded API-based data generation framework"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "requests>=2.31.0",
  "PyYAML>=6.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0"
]

[project.scripts]
datafaker = "datafaker.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
